KTS-3083 "Import from Server Location, you get a Fatal Error" Fixed. Added more document object error checking.
Committed By: Kevin Fourie
Reviewed By: Jonnathan Byrne

git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/trunk@8148 c91229c3-7414-0410-bfa2-8a42b809f60b

+++ lib/documentmanagement/documentutil.inc.php
@@ -259,7 +259,9 @@ class KTDocumentUtil {
         $oUploadChannel->sendMessage(new KTUploadGenericMessage(_kt('Storing contents')));
         $res = KTDocumentUtil::storeContents($oDocument, '', $aOptions);
         if (PEAR::isError($res)) {
-        	$oDocument->delete();
+            if (!PEAR::isError($oDocument)) {
+        	    $oDocument->delete();
+            }
         	return $res;
         }
